--- BurnOutSharp/ProtectionType/TZCopyProtector.cs
@@ -1,32 +1,0 @@
-﻿using System.Collections.Concurrent;
-using System.Collections.Generic;
-using BurnOutSharp.Interfaces;
-using BurnOutSharp.Matching;
-
-namespace BurnOutSharp.ProtectionType
-{
-    public class TZCopyProtector : IPathCheck
-    {
-        /// <inheritdoc/>
-        public ConcurrentQueue<string> CheckDirectoryPath(string path, IEnumerable<string> files)
-        {
-            var matchers = new List<PathMatchSet>
-            {
-                new PathMatchSet(new PathMatch("_742893.016", useEndsWith: true), "TZCopyProtector (Unconfirmed - Please report to us on Github)"),
-            };
-
-            return MatchUtil.GetAllMatches(files, matchers, any: true);
-        }
-
-        /// <inheritdoc/>
-        public string CheckFilePath(string path)
-        {
-            var matchers = new List<PathMatchSet>
-            {
-                new PathMatchSet(new PathMatch("_742893.016", useEndsWith: true), "TZCopyProtector (Unconfirmed - Please report to us on Github)"),
-            };
-
-            return MatchUtil.GetFirstMatch(path, matchers, any: true);
-        }
-    }
-}
